requirements.txt
alembic==1.8.0
anyio==3.6.1
appdirs==1.4.3
apt-clone==0.2.1
apturl==0.5.2
bcrypt==3.2.2
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
ccsm==0.9.14.1
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.1.3
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
entrypoints==0.3
fastapi==0.78.0
filelock==3.7.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
grpcio==1.16.1
gunicorn==20.1.0
h11==0.9.0
httplib2==0.14.0
httptools==0.1.2
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
importlib-metadata==4.11.4
importlib-resources==5.7.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
Mako==1.1.0
MarkupSafe==2.1.1
more-itertools==4.2.0
netaddr==0.7.19
netifaces==0.10.4
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
platformdirs==2.5.2
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.3
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydantic==1.9.1
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==1.9.5
speedtest-cli==2.1.2
SQLAlchemy==1.4.37
starlette==0.19.1
systemd-python==234
tk==0.1.0
tldextract==2.2.1
typing_extensions==4.2.0
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.16.0
virtualenv==20.14.1
virtualenv-clone==0.3.0
wadllib==1.3.3
websockets==8.1
Werkzeug==2.1.2
wsproto==0.15.0
WTForms==3.0.1
xkit==0.0.0
youtube-dl==2021.4.26
zipp==3.8.0
